Adjust the sort index correctly after row removal.
The way the sort index was adjusted when rows were removed was incorrect
in some cases which caused various side-effects.

A demo of this behaviour can be seen in this fiddle: http://jsfiddle.net/einarj/Lj3xagn3/

We fixed the issue by correcting the adjustment algorithm in the
columnSorting plugin.